refactor(#10): Token 발급 로직 수정
