+++ android/app/src/main/java/com/takseha/danari/data/token/TokenInterceptor.kt
@@ -1,6 +1,5 @@
 package com.takseha.danari.data.token
 
-import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.Request
 import okhttp3.Response
@@ -24,7 +23,7 @@ class TokenInterceptor(
 
     private fun addTokenToRequest(request: Request, token: String?): Request {
         return request.newBuilder()
-            .header("Authorization", "Bearer $token")
+            .header("Authorization", token)
             .build()
     }
 }
