+++ android/app/src/main/java/com/takseha/danari/di/Qualifiers.kt
@@ -1,0 +1,11 @@
+package com.takseha.danari.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class AuthRetrofit
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class DefaultRetrofit
